the great refactoring
also fixes integration tests

also fixes integration tests logging on windows

also makes integration tests just log to the integration test logs instead of a file
